Form Validation implemented

+++ src/components/UI/Register/Register.js
@@ -1,37 +1,66 @@
-import { useRef } from 'react';
+import { useState, useRef, useEffect } from 'react';
 import classes from './Register.module.css';
 import { useNavigate } from 'react-router';
-// import useInput from './../../../CustomHooks/useInput'
+import { inputValidation, validEmail, passwordValidation, errorMessages } from './../../../Utilities/FormValidation'
 
 export default function Register(params) {
     
+    const [ error, setError ] = useState({})
     const navigate = useNavigate()
     const username = useRef('')
     const email = useRef('')
     const password = useRef('')
     const passwordConfirm = useRef('')
-    
+
+    console.log( error);
+    let passwordError = errorMessages.passwordRepeat
+    let passwordRepeatError = errorMessages.notValidPassword
+    let emailError = errorMessages.email
+    let emptyFormError = errorMessages.formFields
+
+    // console.log(passwordRepeatError, emailError, emptyFormError);
+    const data = {
+        username: username.current.value,
+        email: email.current.value,
+        password: password.current.value,
+        passwordConfirm: passwordConfirm.current.value
+    }
+
     const signUp = (e) => {
         e.preventDefault()
         
-        if (username.current.value === '' || email.current.value  === '' || password.current.value  === '' || passwordConfirm.current.value === '') {
-            alert('you must feel the form')
-        } 
-        if (password.current.value === passwordConfirm.current.value) {
-            const user = JSON.stringify({
-                username: username.current.value,
-                email: email.current.value,
-                password: password.current.value,
-                passwordConfirm: passwordConfirm.current.value
-            })
+        if (inputValidation(data)) {
 
-            localStorage.setItem('user', user)
-
-            navigate('/')
-
+            if (validEmail(email.current.value)) {
+            
+                if (password.current.value === passwordConfirm.current.value && passwordValidation(password.current.value)) {
+                
+                    const user = JSON.stringify({ data })
+        
+                    localStorage.setItem('user', user)
+                    navigate('/')
+                    
+                } else {
+                    setError({
+                        ...error, 
+                        passwordError     
+                    })  
+                }
+                
+            } else {
+                    setError({
+                        ...error,
+                        emailError
+                    })
+                }
+        
         } else {
-          alert('password must be the same')
+            setError({
+                ...error, 
+                emptyFormError
+            })
         }
+
     }
     return (
         <form onSubmit={signUp}>
@@ -40,23 +69,35 @@ export default function Register(params) {
             </div>
             <div className={`${classes['input-sign']}`}>
                 <label htmlFor="username" className={`#`}>Username</label>
-                <input type="text" id='username' ref={username} />
+                <input type="text" id='username' ref={username} className={`${error.emptyFormError && classes['input-red']}`}/>
             </div>
-            <div className={`${classes['input-sign']}`}>
+            <div className={`${classes['input-sign']} `}>
                 <label htmlFor="userEmail">Email</label>
-                <input type="email" id='userEmail' ref={email}/>
+                <input type="email" id='userEmail' ref={email} className={`${error.emptyFormError && classes['input-red']}`}/>
+                {
+                  error.emailError && <p className={`${classes['error-message']}`}>{errorMessages.email}</p> 
+                }
             </div>
             <div className={`${classes['input-sign']}`}>
                 <label htmlFor="userPassword">Password</label>
-                <input type="password" id='userPassword' ref={password}/>
+                <input type="password" id='userPassword' ref={password} className={`${error.emptyFormError && classes['input-red'] }`}/>
             </div>
+                {
+                  error.passwordError && <p className={`${classes['error-message']}`}>{errorMessages.passwordRepeat}</p> 
+                }
+                {/* {
+                  error.passwordRepeatError && <p className={`${classes['error-message']}`}>{errorMessages.passwordRepeat}</p> 
+                } */}
             <div className={`${classes['input-sign']}`}>
                 <label htmlFor="userPasswordConfirm">Confirm Password</label>
-                <input type="password" id='userPasswordConfirm' ref={passwordConfirm}/>
+                <input type="password" id='userPasswordConfirm' ref={passwordConfirm} className={`${error.emptyFormError && classes['input-red']}`}/>
             </div>
             <div>
                 <button className={`${classes['btn-submit']}`}>Submit</button>
             </div>
+                {
+                  error.emptyFormError && <p className={`${classes['error-message']}`}>{errorMessages.formFields}</p> 
+                }
         </form>
     );
 }
